client: pass a spec to the multi client instead of an expanded list of client ids
The list of client ids is static and will not update over time.
Hence the multi client that is initialized with a fixed list will
not be resilient to changing providers. The fix is to give the multi client
the specs so that it can evaluate matches at each enumeration step.

Also clean up superfluous code in device and connect view controler.

fix https://github.com/bringyour/bringyour/issues/33